+++ Magic/src/main/java/com/elmakers/mine/bukkit/magic/MageProperties.java
@@ -59,7 +59,7 @@ public class MageProperties extends CasterProperties {
             if (overriddenProperties == null) {
                 overriddenProperties = new HashSet<>();
             }
-            overriddenProperties.addAll(activeClass.getEffectiveConfiguration(false).getKeys(false));
+            overriddenProperties.addAll(activeClass.getEffectiveConfiguration().getKeys(false));
         }
 
         super.describe(sender, ignoreProperties, overriddenProperties);